Moved paginator code to a separate module, so it is independent of pages and context files
